requirements.txt
Django==2.2.10
mysqlclient==1.4.6
django-allauth~=0.41.0
django-widget-tweaks~=1.4.8
Pillow~=2.2.1
django-environ~=0.4.5
gunicorn~=20.0.4